flasher-web/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
cryptography==44.0.1
pyserial==3.5
httpx==0.28.1
tinytuya==1.16.1
python-multipart==0.0.20
